Name: VectorEnginePlugin.debugHooks-dtl.5 Author: dtl Time: 5 April 2021, 5:21:20.305 pm UUID: e9d2a282-3c07-4bbc-9cb9-1d923e6d51b6 Ancestors: VectorEnginePlugin.debugHooks-dtl.4
Disable debugging output to the console by default. To enable console output, edit VectorEnginePlugin class>>initialize and set CONSOLE_DEBUG to 1.

Note that time stamps in the console output are very approximate, as most execution time is likely to be consumed in the console output itself. When disabled, there is no overhead because the calls to the output method are guarded by #ifdef.